experiments/method1/requirements.txt
torch
numpy
scikit-learn
bitsandbytes
transformers
peft
accelerate
trl
datasets
evaluate
mypy
black
ipykernel
scikit-multilearn
python-dotenv